requirements.txt
SQLAlchemy
pymysql
python-dotenv
cryptography
Flask-RESTful
Flask-SQLAlchemy